[42182369] Add parameter "address" to `tickets register ...` (#74)
* [42182369] Add parameter "address" to `tickets register ...`

- added "address" parameter to all "tickets register..." commands
- use only utxos/funds from the specified transparent address to fund the ticket registration,
  change is returned to the funding address
  - changed help for all 'tickets register...' commands
  - added support of optional fundingAddress parameter
  - handling optional "funding address" parameter in CreateP2FMSTransactionWithExtra
- moved all ticket register rpc commands to a separate file mnode/rpc/tickets-register.cpp
- google tests:
  - added tests for wallet_ismine.cpp
- python tests:
  - mn_ticket_username_change.py: added tests for registration with address parameter
- rpc-tests.sh: show script execution statistics [11/50|Succeeded:M|Failed:N]

* fixes

* compilation fixes

* fix for pastelid registration

+++ src/rpc/net.cpp
@@ -423,7 +423,7 @@ UniValue getdeprecationinfo(const UniValue& params, bool fHelp)
     obj.pushKV("version", CLIENT_VERSION);
     obj.pushKV("subversion",
         FormatSubVersion(CLIENT_NAME, CLIENT_VERSION, std::vector<string>()));
-    obj.pushKV("deprecationheight", DEPRECATION_HEIGHT);
+    obj.pushKV("deprecationheight", static_cast<uint64_t>(DEPRECATION_HEIGHT));
 
     return obj;
 }
